+++ altrpnjs/app/Customizer/Nodes/MapperNode.ts
@@ -7,8 +7,8 @@ export default class MapperNode extends BaseNode implements NodeInterface
 
   public getContent(): string
   {
-    console.log(this);
-    console.log(this.getChildren());
+    // console.log(this);
+    // console.log(this.getChildren());
     return ''
   }
 
@@ -16,11 +16,16 @@ export default class MapperNode extends BaseNode implements NodeInterface
   {
 
 
-    const {
+    let {
       source = '',
       store = ''
     } = this.data.data;
-
+    if(store.indexOf('context.') == 0){
+      store = store.replace('context.', '')
+    }
+    if(source.indexOf('context.') == 0){
+      source = source.replace('context.', '')
+    }
     // console.log(this.customizer.parsed_data);
     const edges = BaseNode.getEdgesBySourceId(this.data.id, this.customizer.parsed_data)
     //const children = BaseNode.getChildrenBySourceId(this.data.id, this.customizer.parsed_data)
@@ -58,22 +63,35 @@ export default class MapperNode extends BaseNode implements NodeInterface
     if(! source){
       return after
     }
-
-    let result = `
-      await Promise.all(_.map(this.getCustomizerData('context.${source}'),
-        async (item, idx)=>{
-          this.setCustomizerData('context.currentItem', item);
-          this.setCustomizerData('context.currentIndex', idx);
-            ${body}
-        }
-      ))
-    `
+    let result = ''
     if(store){
-      result = `
-    this.setCustomizerData('context.${store}',
-        ${result});
+      result += `
+    this.setCustomizerData('context.${store}', []);
     `
     }
+     result += `
+
+    for(const _idx in this.getCustomizerData('context.${source}', [])){
+      if(this.getCustomizerData('context.${source}').hasOwnProperty(_idx)){
+        this.setCustomizerData('context.currentItem', this.getCustomizerData('context.${source}')[_idx]);
+        this.setCustomizerData('context.currentIndex', _idx);
+        try{
+          ${store ? `
+          this.getCustomizerData('context.${store}').push(
+          await (async ()=>{
+              ${body}
+          })()
+          );` : `
+          await (async ()=>{
+              ${body}
+          })()`}
+        } catch(e){
+          console.error(e)
+        }
+      }
+    }
+
+    `
     return result + after
   }
 }