+++ altrpnjs/helpers/string/stripHtmlTags.ts
@@ -1,0 +1,3 @@
+export default function stripHtmlTags(str) {
+  return str.replace(/<\/?[^>]+(>|$)/g, '');
+}