+++ resources/modules/editor/src/js/components/widgets/InputDateRangeWidget.js
@@ -1,8 +1,12 @@
 import React from "react";
 import moment from 'moment';
 import { DateRangeInput, TimePrecision } from "@blueprintjs/datetime";
+import * as luxon from 'luxon'
 import {changeFormFieldValue} from "../../../../../front-app/src/js/store/forms-data-storage/actions";
+import getDataByPath from "../../../../../front-app/src/js/functions/getDataByPath";
 import isEditor from "../../../../../front-app/src/js/functions/isEditor";
+import '../../../sass/blueprint-datetime.scss'
+const {DateTime} = luxon
 
 class InputDateRangeWidget extends Component {
   constructor(props){
@@ -35,17 +39,41 @@ class InputDateRangeWidget extends Component {
         this.timePrecision = TimePrecision.MINUTE
         break
     }
+    let content_start_default = props.element.getSettings('content_default_value_start')
 
-    this.handleChange = this.handleChange.bind(this)
+    let content_end_default = props.element.getSettings('content_default_value_end')
+    if(content_start_default && content_start_default.includes('{{')){
+      content_start_default = getDataByPath(content_start_default.replace('{{', '').replace('}}', ''))
+    }
+
+    if(content_start_default){
+      content_start_default = DateTime.fromSQL(content_start_default).toJSDate()
+    }
+    if(content_end_default && content_end_default.includes('{{')){
+      content_end_default = getDataByPath(content_end_default.replace('{{', '').replace('}}', ''))
+    }
+    if(content_end_default){
+      content_end_default = DateTime.fromSQL(content_end_default).toJSDate()
+    }
+
+    if(!isEditor()) {
+      if(content_start_default){
+        this.dispatchFieldValueToStore(content_start_default, 0, true)
+      }
+      if(content_end_default){
+        this.dispatchFieldValueToStore(content_end_default, 1, true)
+      }
+    }
   }
 
-  handleChange(value) {
-    if(isEditor()) {
-      this.setState(s => ({
-        ...s,
-        value
-      }))
-    } else {
+  handleChange = (value)=> {
+
+    this.setState(s => ({
+      ...s,
+      value
+    }))
+    if(!isEditor()) {
+
       this.dispatchFieldValueToStore(value[0], 0, true)
       this.dispatchFieldValueToStore(value[1], 1, true)
     }
@@ -61,7 +89,6 @@ class InputDateRangeWidget extends Component {
 
     let formId;
     let fieldName;
-
     if(index === 0) {
       formId = this.props.element.getFormId("form_id_start");
       fieldName = this.props.element.getFieldId("field_id_start");
@@ -73,27 +100,39 @@ class InputDateRangeWidget extends Component {
       fieldName = replaceContentWithData(fieldName);
     }
     if (_.isObject(this.props.appStore) && fieldName && formId) {
+      if(value){
+        value = DateTime.fromJSDate(value).toSQL()
+      }
+
       this.props.appStore.dispatch(
         changeFormFieldValue(fieldName, value, formId, userInput)
       );
-      // if (userInput) {
-      //   const change_actions = this.props.element.getLockedSettings("change_actions");
-      //
-      //   if (change_actions && !isEditor()) {
-      //     const actionsManager = (
-      //       await import(
-      //         /* webpackChunkName: 'ActionsManager' */
-      //         "../../../../../front-app/src/js/classes/modules/ActionsManager.js"
-      //         )
-      //     ).default;
-      //     await actionsManager.callAllWidgetActions(
-      //       this.props.element.getIdForAction(),
-      //       "change",
-      //       change_actions,
-      //       this.props.element
-      //     );
-      //   }
-      // }
+
+      let query_sync = this.props.element.getLockedSettings(
+        "query_sync"
+      );
+      if(!isEditor() && query_sync){
+        const updateQueryString = (await import('../../../../../front-app/src/js/functions/updateQueryString')).default
+        updateQueryString(fieldName, value)
+
+        if (userInput) {
+          const change_actions = this.props.element.getLockedSettings("change_actions");
+          if (change_actions && !isEditor()) {
+            const actionsManager = (
+              await import(
+                /* webpackChunkName: 'ActionsManager' */
+                "../../../../../front-app/src/js/classes/modules/ActionsManager.js"
+                )
+            ).default;
+            await actionsManager.callAllWidgetActions(
+              this.props.element.getIdForAction(),
+              "change",
+              change_actions,
+              this.props.element
+            );
+          }
+        }
+      }
     }
   };
 
@@ -115,11 +154,16 @@ class InputDateRangeWidget extends Component {
 
       valueStart = _.get(appStore.getState().formsStore, `${formIdStart}`, '')
       valueStart = valueStart[fieldNameStart] || null
+      if(valueStart){
+        valueStart = DateTime.fromSQL(valueStart).toJSDate()
+      }
       let valueEnd
 
       valueEnd = _.get(appStore.getState().formsStore, `${formIdEnd}`, '')
       valueEnd = valueEnd[fieldNameEnd] || null
-
+      if(valueEnd){
+        valueEnd = DateTime.fromSQL(valueEnd).toJSDate()
+      }
       value = [
         valueStart,
         valueEnd
@@ -145,10 +189,19 @@ class InputDateRangeWidget extends Component {
   }
 
   render(){
+    const {element} = this.props
+    const format = element.getLockedSettings('content_format') || 'YYYY-MM-DD';
 
-    const format = this.props.element.getLockedSettings('content_format') || 'YYYY-MM-DD';
-    const dayPickerProps = {};
 
+    const dayPickerProps = {
+
+    };
+    if(getDataByPath(this.getContent('max_date'))){
+      dayPickerProps.maxDate = DateTime.fromSQL(getDataByPath(this.getContent('max_date'))).toJSDate()
+    }
+    if(getDataByPath(this.getContent('min_date'))){
+      dayPickerProps.minDate = DateTime.fromSQL(getDataByPath(this.getContent('min_date'))).toJSDate()
+    }
     let value = this.getValue();
 
     const locale = this.locale || "ru";
@@ -165,25 +218,27 @@ class InputDateRangeWidget extends Component {
       frame = document.getElementById("editorContent").contentWindow.document.body
     }
     let classes =
-      this.getClasses() + (this.props.element.getResponsiveLockedSetting('position_css_classes') || "")
+      this.getClasses() + (element.getResponsiveLockedSetting('position_css_classes') || "")
 
     return <DateRangeInput
       className={classes}
-      minDate={new Date(1900, 1, 1)}
       formatDate={date => moment(date).locale(locale).format(format)}
       parseDate={str => moment(str, this.typeDate).locale(locale).toDate()}
       onChange={this.handleChange}
+      maxDate={dayPickerProps.maxDate}
+      minDate={dayPickerProps.minDate}
+      allowSingleDayRange={element.getSettings("allow_single_day_range")}
       dayPickerProps={dayPickerProps}
       value={value}
       startInputProps={{
-        placeholder: this.props.element.getResponsiveLockedSetting("start_placeholder", "", "start date")
+        placeholder: element.getResponsiveLockedSetting("start_placeholder", "", "start date")
       }}
       endInputProps={{
-        placeholder: this.props.element.getResponsiveLockedSetting("end_placeholder", "", "end date")
+        placeholder: element.getResponsiveLockedSetting("end_placeholder", "", "end date")
       }}
       popoverProps={{
         usePortal: true,
-        popoverClassName: "altrp-portal" + this.props.element.getId(),
+        popoverClassName: "altrp-portal" + element.getId(),
         portalContainer: frame
       }}