+++ resources/modules/editor/src/js/components/widgets/InputRadioWidget.js
@@ -1,9 +1,7 @@
 import isEditor from "../../../../../front-app/src/js/functions/isEditor";
-import convertData from "../../../../../front-app/src/js/functions/convertData";
 import parseOptionsFromSettings from "../../../../../front-app/src/js/functions/parseOptionsFromSettings";
 import parseParamsFromString from "../../../../../front-app/src/js/functions/parseParamsFromString";
 import replaceContentWithData from "../../../../../front-app/src/js/functions/replaceContentWithData";
-import getDataByPath from "../../../../../front-app/src/js/functions/getDataByPath";
 import getDataFromLocalStorage from "../../../../../front-app/src/js/functions/getDataFromLocalStorage";
 import renderAssetIcon from "../../../../../front-app/src/js/functions/renderAssetIcon";
 import altrpCompare from "../../../../../front-app/src/js/functions/altrpCompare";
@@ -601,7 +599,7 @@ class InputRadioWidget extends Component {
             let checked;
 
             checked = altrpCompare(value, option.value, "==");
-            console.log(checked);
+
             return (
               <Radio
                 className={`${classes} altrp-field-radio ${checked ? "active" : ""} ${radioPosition == 'right' ? 'bp3-align-right' : ''}`}